Backend/requirements.txt
Flask
flask-cors
Flask-Session
pymongo
python-dotenv
Pillow
transformers
torch
google-generativeai
